extended container classes, implemented simple updater and NakedSingleMethod

+++ src/ch/fhnw/ip5/sudoku/sudoku/Column.java
@@ -4,11 +4,11 @@ public class Column {
 	
 	private Cell[] cells;
 	
-	public Column(int height) {
+	public Column(byte height) {
 		cells = new Cell[height];
 	}
 	
-	public void setCell(Cell cell, int pos) {
+	public void setCell(Cell cell, byte pos) {
 		cells[pos] = cell;
 	}
